+++ public/js/src/module/connection.js
@@ -14,6 +14,7 @@ var CONNECTION_URL = '/connection';
 // location.search is added to pass the lang= parameter, in case this is used to override browser/system locale
 var TRANSFORM_URL = '/transform/xform' + location.search;
 var TRANSFORM_HASH_URL = '/transform/xform/hash';
+var EXPORT_URL = '/export/get-url';
 var SUBMISSION_URL = ( settings.enketoId ) ? '/submission/' + settings.enketoIdPrefix + settings.enketoId + location.search : null;
 var INSTANCE_URL = ( settings.enketoId ) ? '/submission/' + settings.enketoIdPrefix + settings.enketoId : null;
 var MAX_SIZE_URL = ( settings.enketoId ) ? '/submission/max-size/' + settings.enketoIdPrefix + settings.enketoId : null;
@@ -95,6 +96,28 @@ function uploadRecord( record ) {
             console.debug( 'results of all batches submitted', results );
             return results[ 0 ];
         } );
+}
+
+function getDownloadUrl( zipFile ) {
+    return new Promise( function( resolve, reject ) {
+        var formData = new FormData();
+        formData.append( 'export', zipFile, zipFile.name );
+
+        $.ajax( EXPORT_URL, {
+                type: 'POST',
+                data: formData,
+                cache: false,
+                contentType: false,
+                processData: false
+            } )
+            .done( function( data ) {
+                resolve( data.downloadUrl );
+            } )
+            .fail( function( jqXHR, textStatus ) {
+                console.error( jqXHR, textStatus );
+                reject( new Error( textStatus || 'Failed to connect with Enketo server.' ) );
+            } );
+    } );
 }
 
 /**
@@ -512,5 +535,6 @@ module.exports = {
     getFormPartsHash: getFormPartsHash,
     getMediaFile: getMediaFile,
     getExistingInstance: getExistingInstance,
-    getManifestVersion: getManifestVersion
+    getManifestVersion: getManifestVersion,
+    getDownloadUrl: getDownloadUrl
 };