+++ public/js/src/module/store.js
@@ -482,6 +482,7 @@ recordStore = {
                 name: record.name,
                 xml: record.xml,
                 files: fileKeys,
+                created: new Date().getTime(),
                 updated: new Date().getTime(),
                 draft: record.draft
             } )
@@ -542,6 +543,7 @@ recordStore = {
                     name: record.name,
                     xml: record.xml,
                     files: fileKeys,
+                    created: ( result && result.created ? result.created : new Date().getTime() ),
                     updated: new Date().getTime(),
                     draft: record.draft
                 } );