+++ public/js/src/module/utils.js
@@ -33,6 +33,33 @@ function blobToDataUri( blob ) {
 }
 
 /**
+ * Converts a Blob to a an ArrayBuffer
+ *
+ * @param  {Blob} blob The blob
+ * @return {Promise}
+ */
+function blobToArrayBuffer( blob ) {
+    var reader = new window.FileReader();
+
+    return new Promise( function( resolve, reject ) {
+        reader.onloadend = function() {
+            resolve( reader.result );
+        };
+        reader.onerror = function( e ) {
+            reject( e );
+        };
+
+        // There is some quirky Chrome and Safari behaviour if blob is undefined or a string
+        // so we peform an additional check
+        if ( !( blob instanceof Blob ) ) {
+            reject( new Error( 'TypeError: Require Blob' ) );
+        } else {
+            reader.readAsArrayBuffer( blob );
+        }
+    } );
+}
+
+/**
  * The inverse of blobToDataUri, that converts a dataURL back to a Blob
  *
  * @param  {string} dataURI dataURI
@@ -130,6 +157,7 @@ function _throwInvalidXmlNodeName( name ) {
 
 module.exports = {
     blobToDataUri: blobToDataUri,
+    blobToArrayBuffer: blobToArrayBuffer,
     dataUriToBlob: dataUriToBlob,
     getThemeFromFormStr: getThemeFromFormStr,
     getTitleFromFormStr: getTitleFromFormStr,