+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased] 
+---------------------
+##### Added
+- Export functionality.
+
 [1.21.2] - 2015-12-18
 ---------------------
 ##### Changed